Feature event join

+++ src/Controller/EventController.php
@@ -6,9 +6,11 @@ use App\Entity\Category;
 use App\Entity\Event;
 use App\Form\EventType;
 use App\Data\SearchData;
+use App\Entity\Attendant;
 use App\Form\SearchFormType;
 use App\Repository\EventRepository;
 use App\Repository\AttendantRepository;
+use App\Service\isAttendant;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -191,11 +193,81 @@ class EventController extends AbstractController
 		}
 
 		/**
+		 * To join an event
+		 *
+		 * @Route("/event/{id<\d+>}/join", name="app_event_join", methods={"GET"})
+		 */
+		public function join(Event $event, EntityManagerInterface $em, Request $request, AttendantRepository $ar, isAttendant $checkAttendant) {
+
+			if(null === $event) {
+				throw $this->createNotFoundException("404 - Sortie introuvable");
+			}
+
+			$user = $this->getUser();
+
+			//! Vérifier si il y a une place disponible
+			// je récupère le max des participants
+			$maxAttendant = $event->getMaxAttendants();
+			// je récupère le nbr des participants actuels
+			$nbrAttendants = count($event->getAttendants());
+
+			//! Vérifier si l'utilisateur n'est pas déjà participant
+			// je récupère la liste des participants de la sortie sélectionnée
+			$attendantList = $ar->findByEvent($event);
+
+			// je fais appel à mon service qui me permet de vérifier si
+			// l'utilisateur participe déjà à cette sortie
+			$isAttendant = $checkAttendant->checkIsAttendant($attendantList, $user);
+
+			//! Vérifier si l'utilisateur est connecté
+			if($user) {
+				// On vérifie si il y a des places de disponible
+				if($nbrAttendants < $maxAttendant) {
+
+					// Si il n'est pas déjà un participant
+					if($isAttendant === false) {
+						$attendant = new Attendant();
+
+						$attendant->setUser($user);
+						$attendant->setEvent($event);
+
+						$em->persist($attendant);
+						$em->flush();
+
+						// Flash message
+						$this->addFlash('success', 'Vous avez bien été ajouté à la sortie '.$event->getTitle().' !');
+
+						// Redirection sur la page de l'évènement correspondant
+						return $this->redirectToRoute('app_event_show', [
+							'id' => $event->getId(),
+						]);
+					}
+					// on indique à l'utilisateur qu'il participe déjà à cette sortie
+					$this->addFlash('warning', 'Vous êtes déjà inscrit à cette sortie !');
+
+					return $this->redirect($request->headers->get('referer'));
+
+				}
+
+				// on indique à l'utilisateur qu'il n'y a plus de places disponibles
+				$this->addFlash('danger', 'Désolé il n\'y a plus de places disponibles sur cette sortie !');
+
+				return $this->redirect($request->headers->get('referer'));
+			
+			}
+
+			// on lui indique qu'il est nécessaire de se connecter pour participer à une sortie
+			$this->addFlash('danger', 'Connectez vous pour participer à une sortie !');
+
+			return $this->redirectToRoute('app_login');
+		}
+
+		/**
 		 * To leave an event
 		 *
 		 * @Route("/event/{id<\d+>}/leave", name="app_event_leave", methods={"GET"})
 		 */
-		public function leave(Event $event = null, AttendantRepository $ar, EntityManagerInterface $em, Request $request)
+		public function leave(Event $event = null, AttendantRepository $ar, EntityManagerInterface $em, Request $request, isAttendant $checkAttendant)
 		{
 
 			$user = $this->getUser();
@@ -208,14 +280,35 @@ class EventController extends AbstractController
 			// requete custom pour la récupération
 			$attendant = $ar->findByUserEvent($user, $event);
 
-			// suppresion de la participation à la sortie de la bdd
-			$em->remove($attendant[0]);
-			$em->flush();
+			//! Vérifier si l'utilisateur est un participant
+			// je récupère la liste des participants de la sortie sélectionnée
+			$attendantList = $ar->findByEvent($event);
 
-			// Flash message
-			$this->addFlash('success', 'Vous avez quitté la sortie avec succès !');
+			// je fais appel à mon service qui me permet de vérifier si
+			// l'utilisateur participe déjà à cette sortie
+			$isAttendant = $checkAttendant->checkIsAttendant($attendantList, $user);
 
-			// redirection dans la page courante
+			if($user) {
+
+				if($isAttendant) {
+					// suppresion de la participation à la sortie de la bdd
+					$em->remove($attendant[0]);
+					$em->flush();
+
+					// Flash message
+					$this->addFlash('success', 'Vous avez quitté la sortie avec succès !');
+
+					// redirection dans la page courante
+					return $this->redirect($request->headers->get('referer'));
+				}
+				// on indique à l'utilisateur qu'il ne participe pas à cette sortie
+				$this->addFlash('warning', 'Vous n\'êtes pas inscrit à cette sortie !');
+
+				return $this->redirect($request->headers->get('referer'));
+			}
+			// on indique à l'utilisateur qu'il doit se connecter
+			$this->addFlash('danger', 'Vous devez vous connecter pour effetuer cette action !');
+
 			return $this->redirect($request->headers->get('referer'));
 		}
 }
